Merge server and cli modules
Combine the CLI and the server into a single module. Intent is to have a single module that is the core program and its primary cli and have other modules be basises for plugins that are discovered at runtime, such that within this repository there are no modules that depend on each other.